Asistente Virtual AI Ubimia/requirements.txt
streamlit==1.39.0
pandas==2.2.2
numpy==1.26.4
python-docx==1.1.2
python-pptx==0.6.23
matplotlib==3.8.4
requests==2.32.3
beautifulsoup4==4.12.3
openpyxl==3.1.5
openai>=1.52.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
jinja2>=3.1.4
webvtt-py>=0.5.1
srt>=3.5.3
pypandoc>=1.13
rapidfuzz>=3.9.6
PyYAML==6.0.2
pyyaml
